v1.3.0

+++ Projects/STM324x9I_EVAL/Applications/LwIP/LwIP_HTTP_Server_Netconn_RTOS/Inc/main.h
@@ -1,20 +1,30 @@
 /**
   ******************************************************************************
-  * @file    LwIP/LwIP_HTTP_Server_Netconn_RTOS/Inc/main.h 
+  * @file   LwIP\LwIP_HTTP_Server_Netconn_RTOS\Inc\main.h 
   * @author  MCD Application Team
+  * @version V1.1.0
+  * @date    26-June-2014
   * @brief   Header for main.c module
   ******************************************************************************
   * @attention
   *
-  * Copyright (c) 2017 STMicroelectronics.
-  * All rights reserved.
+  * <h2><center>&copy; COPYRIGHT(c) 2014 STMicroelectronics</center></h2>
   *
-  * This software is licensed under terms that can be found in the LICENSE file
-  * in the root directory of this software component.
-  * If no LICENSE file comes with this software, it is provided AS-IS.
+  * Licensed under MCD-ST Liberty SW License Agreement V2, (the "License");
+  * You may not use this file except in compliance with the License.
+  * You may obtain a copy of the License at:
+  *
+  *        http://www.st.com/software_license_agreement_liberty_v2
+  *
+  * Unless required by applicable law or agreed to in writing, software 
+  * distributed under the License is distributed on an "AS IS" BASIS, 
+  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  * See the License for the specific language governing permissions and
+  * limitations under the License.
   *
   ******************************************************************************
   */
+
 /* Define to prevent recursive inclusion -------------------------------------*/
 #ifndef __MAIN_H
 #define __MAIN_H
@@ -26,32 +36,34 @@
 /* Includes ------------------------------------------------------------------*/
 #include "stm32f4xx_hal.h"
 #include "stm324x9i_eval.h"
-#include "cmsis_os.h"
+#include "stm324x9i_eval_lcd.h"
+#include "stm324x9i_eval_io.h"
 
 /* Exported types ------------------------------------------------------------*/
 /* Exported constants --------------------------------------------------------*/
-#define USE_LCD        /* enable LCD  */
-
+#define USE_LCD        /* enable LCD  */  
+//#define USE_DHCP       /* enable DHCP, if disabled static address is used*/
+ 
 /*Static IP ADDRESS*/
-#define IP_ADDR0   ((uint8_t)192U)
-#define IP_ADDR1   ((uint8_t)168U)
-#define IP_ADDR2   ((uint8_t)0U)
-#define IP_ADDR3   ((uint8_t)10U)
-
+#define IP_ADDR0   192
+#define IP_ADDR1   168
+#define IP_ADDR2   0
+#define IP_ADDR3   10
+   
 /*NETMASK*/
-#define NETMASK_ADDR0   ((uint8_t)255U)
-#define NETMASK_ADDR1   ((uint8_t)255U)
-#define NETMASK_ADDR2   ((uint8_t)255U)
-#define NETMASK_ADDR3   ((uint8_t)0U)
+#define NETMASK_ADDR0   255
+#define NETMASK_ADDR1   255
+#define NETMASK_ADDR2   255
+#define NETMASK_ADDR3   0
 
 /*Gateway Address*/
-#define GW_ADDR0   ((uint8_t)192U)
-#define GW_ADDR1   ((uint8_t)168U)
-#define GW_ADDR2   ((uint8_t)0U)
-#define GW_ADDR3   ((uint8_t)1U)
+#define GW_ADDR0   192
+#define GW_ADDR1   168
+#define GW_ADDR2   0
+#define GW_ADDR3   1 
 
 /* Exported macro ------------------------------------------------------------*/
-/* Exported functions ------------------------------------------------------- */
+/* Exported functions ------------------------------------------------------- */  
 
 
 #ifdef __cplusplus
@@ -59,3 +71,6 @@
 #endif
 
 #endif /* __MAIN_H */
+
+
+/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
